// src/shared/types/blockTypes.ts
export enum BlockTypes {
   Unstyled = "unstyled",
   HeaderOne = "header-one",
   HeaderTwo = "header-two",
   HeaderThree = "header-three",
   HeaderFour = "header-four",
   HeaderFive = "header-five",
   HeaderSix = "header-six",
   UnorderedListItem = "unordered-list-item",
   OrderedListItem = "ordered-list-item",
   BlockQuote = "blockQuote",
   CodeBlock = "code-Block",
   Atomic = "atomic",
}